requirements.txt

streamlit
langgraph
langchain
langchain-core
langchain-google-genai
imap-tools
requests
google-auth
google-auth-oauthlib
plotly
